# auth-server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "description": "My API description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "components": {
    "securitySchemes": {
      "jwt": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
      "userCredentials": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "securepassword123"
          }
        },
        "required": ["email", "password"]
      },
      "workout": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Late night workout"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/exercise"
            }
          }
        }
      },
      "exercise": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "number",
            "example": 43 
          },
          "name": {
            "type": "string",
            "example": "push ups"
          },
          "orderNumber": {
            "type": "number",
            "example": 1
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/exerciseSet"
            }
          }
        }
      },
      "exerciseSet": {
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "number",
            "example": 1
          },
          "weight": {
            "type": "number",
            "example": 45
          },
          "reps": {
            "type": "number",
            "example": 12
          },
          "completed": {
            "type": "boolean",
            "example": "true"
          }
        }
      }
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Login to the application",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/verify": {
      "get": {

      }
    },
    "/workouts": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Add new workout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/workout"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/workouts/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/workout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}